requirements.txt
# Core API Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
python-multipart>=0.0.17
python-dotenv>=1.0.1

# HTTP Client
requests>=2.31.0

# AI/ML Services
openai>=1.52.2

# Media Download Tools
yt-dlp>=2024.1.0
gallery-dl>=1.26.0

# Cookie auto-refresh (cron job semanal no VPS)
playwright>=1.44.0
